Clean up `scripts/run_*.py`

+++ scripts/run_multi.py
@@ -1,6 +1,7 @@
 import logging
-import os
 import shutil
+from datetime import datetime
+from pathlib import Path
 
 import ray
 from ray import tune
@@ -15,30 +16,26 @@ logger = logging.getLogger(__name__)
 # ---------- OPTIONS -----------
 LOCAL_MODE = False
 OVERWRITE = True
-RUN_TAG = "test_new_run"
-RUNS_HOME = "/snel/share/runs/lfads-torch/validation"
-RUN_DIR = f"{RUNS_HOME}/multi/{RUN_TAG}"
+RUN_TAG = datetime.now().strftime("%Y%m%d-%H%M%S")
+RUNS_HOME = Path("/snel/share/runs/lfads-torch/validation")
+RUN_DIR = RUNS_HOME / "multi" / RUN_TAG
 # ------------------------------
 
 # Initialize the `ray` server in local mode if necessary
 if LOCAL_MODE:
     ray.init(local_mode=True)
 # Overwrite the directory if necessary
-if os.path.exists(RUN_DIR):
-    if OVERWRITE:
-        logger.warning(f"Overwriting multi-run at {RUN_DIR}")
-        shutil.rmtree(RUN_DIR)
-    else:
-        raise OSError(
-            "The multi-run directory already exists. "
-            "Set `OVERWRITE=True` or create a new `RUN_TAG`."
-        )
+if RUN_DIR.exists() and OVERWRITE:
+    shutil.rmtree(RUN_DIR)
+RUN_DIR.mkdir()
+# Copy this script into the run directory
+shutil.copyfile(__file__, RUN_DIR / Path(__file__).name)
 # Run the hyperparameter search
 tune.run(
     tune.with_parameters(run_model, config_name="multi.yaml"),
     metric="valid/recon_smth",
     mode="min",
-    name=os.path.basename(RUN_DIR),
+    name=RUN_DIR.name,
     config=dict(
         model=dict(
             dropout_rate=tune.uniform(0.0, 0.7),
@@ -52,7 +49,7 @@ tune.run(
     ),
     resources_per_trial=dict(cpu=3, gpu=0.5),
     num_samples=20,
-    local_dir=os.path.dirname(RUN_DIR),
+    local_dir=RUN_DIR.parent,
     search_alg=BasicVariantGenerator(random_state=0),
     scheduler=FIFOScheduler(),
     verbose=1,
