+++ scripts/run_single.py
@@ -1,6 +1,8 @@
 import logging
 import os
 import shutil
+from datetime import datetime
+from pathlib import Path
 
 from lfads_torch.run_model import run_model
 
@@ -8,22 +10,15 @@ logger = logging.getLogger(__name__)
 
 # ---------- OPTIONS -----------
 OVERWRITE = True
-RUN_TAG = "test_new_run"
-RUNS_HOME = "/snel/share/runs/lfads-torch/validation"
-RUN_DIR = f"{RUNS_HOME}/single/{RUN_TAG}"
+RUN_TAG = datetime.now().strftime("%Y%m%d-%H%M%S")
+RUNS_HOME = Path("/snel/share/runs/lfads-torch/validation")
+RUN_DIR = RUNS_HOME / "single" / RUN_TAG
 # ------------------------------
 
 # Overwrite the directory if necessary
-if os.path.exists(RUN_DIR):
-    if OVERWRITE:
-        logger.warning(f"Overwriting single-run at {RUN_DIR}")
-        shutil.rmtree(RUN_DIR)
-    else:
-        raise OSError(
-            "The single-run directory already exists. "
-            "Set `OVERWRITE=True` or create a new `RUN_TAG`."
-        )
+if RUN_DIR.exists() and OVERWRITE:
+    shutil.rmtree(RUN_DIR)
+RUN_DIR.mkdir()
 # Switch to the `RUN_DIR` and train the model
-os.makedirs(RUN_DIR)
 os.chdir(RUN_DIR)
 run_model(config_name="single.yaml")
